+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Upcoming
+## Added
+- Added support for using environmental variable values for the default value
+of options.
 
 # v1.0.1
 ## Bug Fixes